сделал контроллеры для api
